refactor(constants): move constants to specific files and update references in MainViewModel and ContentView
