+++ ConfigForge/Utilities/KubeConfigConstants.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct KubeConfigConstants {
+    static let defaultKubeConfigPath = "~/.kube/config"
+    static let defaultKubeBackupFileName = "kubeconfig_backup"
+}
