+++ ConfigForge/Utilities/SSHConfigConstants.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct SSHConfigConstants {
+    static let defaultSSHConfigPath = "~/.ssh/config"
+    // Default filenames for backup (base names)
+    static let defaultBackupFileName = "ssh_config_backup"
+}
